Refactor(jira): Move some classes into new packages.
