+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/message/IssueContentModel.java
