+++ src/main/java/com/synopsys/integration/alert/channel/jira/common/JiraMessageParser.java
@@ -36,9 +36,9 @@ import com.synopsys.integration.alert.common.channel.message.ChannelMessageParse
 import com.synopsys.integration.alert.common.channel.message.MessageSplitter;
 import com.synopsys.integration.alert.common.message.model.ComponentItem;
 import com.synopsys.integration.alert.common.message.model.LinkableItem;
-import com.synopsys.integration.alert.issuetracker.IssueContentModel;
 import com.synopsys.integration.alert.issuetracker.IssueProperties;
 import com.synopsys.integration.alert.issuetracker.OperationType;
+import com.synopsys.integration.alert.issuetracker.message.IssueContentModel;
 import com.synopsys.integration.alert.issuetracker.message.IssueCreationRequest;
 
 @Component
