+++ src/main/java/com/synopsys/integration/alert/channel/jira/server/actions/JiraServerDistributionTestAction.java
@@ -33,7 +33,7 @@ import com.synopsys.integration.alert.channel.jira.server.JiraServerContextBuild
 import com.synopsys.integration.alert.common.channel.ChannelDistributionTestAction;
 import com.synopsys.integration.alert.common.message.model.MessageResult;
 import com.synopsys.integration.alert.common.persistence.accessor.FieldAccessor;
-import com.synopsys.integration.alert.issuetracker.IssueTrackerContext;
+import com.synopsys.integration.alert.issuetracker.config.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.jira.server.JiraServerCreateIssueTestAction;
 import com.synopsys.integration.alert.issuetracker.jira.server.JiraServerService;
 import com.synopsys.integration.alert.issuetracker.message.IssueTrackerResponse;
